svg: Disable ©®*#0-9 for font selection issues
Multiple Linux programs are parsing the provided fontconfig
incorrectly or the fontconfig is incorrect.

The problem is, probably, related to the definition of the fontconfig
edit mode "prepend_first". Either way, until a solution can be found
these characters and therefore associated ligatures will be disabled.

Fixes #35, fixes #37, fixes #39 and creates #40.