windows: Don't forget current working directory (#65)
When you right click and run the `install.cmd` script as an administrator, the
script forgets the working directory and instead uses `%systemroot%/system32`
as the working directory, this throws an error later on during installation
because we are looking for the file `EmojiOneColor-SVGinOT` in
`%systemroot%/system32`.

The two line additions fix this problem. Now you can right click on the script
and install the fonts from anywhere.

You can reproduce this bug by running the script as an administrator from the
extracted path (right click -> Run as administrator).